+++ Local positioning system/Manager/AlertManager.swift
@@ -1,0 +1,24 @@
+//
+//  File.swift
+//  Local positioning system
+//
+//  Created by Леонид Лукашевич on 14.05.2022.
+//
+
+import UIKit
+
+class AlertManager {
+    static func showAlert(title: String, message: String) {
+        
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        let okAction = UIAlertAction(title: "OK", style: .default)
+        
+        alert.addAction(okAction)
+        
+        let alertWindow = UIWindow(frame: UIScreen.main.bounds)
+        alertWindow.rootViewController = UIViewController()
+        alertWindow.windowLevel = UIWindow.Level.alert + 1
+        alertWindow.makeKeyAndVisible()
+        alertWindow.rootViewController?.present(alert, animated: true, completion: nil)
+    }
+}
