+++ Local positioning system/Manager/GettingInsideManager.swift
@@ -12,58 +12,61 @@ class GettingInsideManager {
     lazy var buildingCoordinates = BuildingCoordinate()
     lazy var buildingArea = BuildingArea()
     
-    lazy var locationServicesManager = LocationServicesManager()
+    lazy var locationServicesManager = LocationServicesManager.shared
     lazy var timerManager = TimerManager()
     
-    func startDetectionGettingInsideArea(completionHandler: @escaping (CLLocationCoordinate2D) -> ()) {
-        timerManager.startTimer(timeInterval: 10) { [weak self] in
+    func startDetectionGettingInside(completionHandler: @escaping (CLLocationCoordinate2D) -> ()) {
+        let completion: ((CLLocation) -> ()) = { [weak self] location in
             guard let self = self else { return }
             
             self.getCurrentUserLocation() { location in
-                let isInside = self.checkGettingInside(in: self.buildingArea, userLocation: location)
-
+                let isInside = self.checkGettingInside(in: self.buildingCoordinates, userLocation: location)
+                
                 if isInside {
-                    self.timerManager.stopTimer()
-                    self.startDetectionGettingInsideBuilding(completionHandler: completionHandler)
+                    completionHandler(location)
+                    self.locationServicesManager.locationManager.stopUpdatingLocation()
                 }
             }
         }
+        
+        setCompletionToLocationManagerDelegate(completion: completion)
+        locationServicesManager.locationManager.startUpdatingLocation()
     }
     
-    func startDetectionGettingInsideBuilding(completionHandler: @escaping (CLLocationCoordinate2D) -> ()) {
-        timerManager.startTimer(timeInterval: 1) { [weak self] in
-            guard let self = self else { return }
-            
-            self.getCurrentUserLocation() { location in
-                let isInArea = self.checkGettingInside(in: self.buildingArea, userLocation: location)
-                if !isInArea {
-                    self.timerManager.stopTimer()
-                    self.startDetectionGettingInsideArea(completionHandler: completionHandler)
-                }
-                
-                let isInside = self.checkGettingInside(in: self.buildingCoordinates, userLocation: location)
-                if isInside {
-                    self.timerManager.stopTimer()
-                    completionHandler(location)
-                    
-//                    let closure: ((CLLocation) -> ()) = { loc in
-//                        completionHandler(loc.coordinate)
+//    func startDetectionGettingInsideBuilding(completionHandler: @escaping (CLLocationCoordinate2D) -> ()) {
+//        timerManager.startTimer(timeInterval: 1) { [weak self] in
+//            guard let self = self else { return }
 //
-//                        if !self.checkGettingInside(in: self.buildingCoordinates, userLocation: loc) {
-//                            self.locationServicesManager.locationManager.stopUpdatingLocation()
-//                            self.positioningManager.positioningMotionManager.stopDeviceMotionUpdate()
-//                            self.startDetectionGettingInsideBuilding()
-//                        }
-//                    }
-                    
-//                    self.setCompletionToLocationManagerDelegate(completion: completionHandler)
-//                    self.locationServicesManager.locationManager.startUpdatingLocation()
-                    
-//                    self.positioningManager.startRecordingMotions(coordinatesOfStart: location, closure: closure)
-                }
-            }
-        }
-    }
+//            self.getCurrentUserLocation() { location in
+//                let isInArea = self.checkGettingInside(in: self.buildingArea, userLocation: location)
+//                if !isInArea {
+//                    self.timerManager.stopTimer()
+//                    self.startDetectionGettingInside(completionHandler: completionHandler)
+//                }
+//
+//                let isInside = self.checkGettingInside(in: self.buildingCoordinates, userLocation: location)
+//                if isInside {
+//                    self.timerManager.stopTimer()
+//
+//
+////                    let closure: ((CLLocation) -> ()) = { loc in
+////                        completionHandler(loc.coordinate)
+////
+////                        if !self.checkGettingInside(in: self.buildingCoordinates, userLocation: loc) {
+////                            self.locationServicesManager.locationManager.stopUpdatingLocation()
+////                            self.positioningManager.positioningMotionManager.stopDeviceMotionUpdate()
+////                            self.startDetectionGettingInsideBuilding()
+////                        }
+////                    }
+//
+////                    self.setCompletionToLocationManagerDelegate(completion: completionHandler)
+////                    self.locationServicesManager.locationManager.startUpdatingLocation()
+//
+////                    self.positioningManager.startRecordingMotions(coordinatesOfStart: location, closure: closure)
+//                }
+//            }
+//        }
+//    }
     
     func checkGettingInside(in coordinates: BuildingProtocol, userLocation: CLLocationCoordinate2D) -> Bool {
         let lb = coordinates.leftBottom
@@ -98,9 +101,9 @@ class GettingInsideManager {
         return (secondX - firstX) * (pointY - firstY) - (secondY - firstY) * (pointX - firstX)
     }
     
-    func stopDetectionGettingInsideArea() {
-        timerManager.stopTimer()
-    }
+//    func stopDetectionGettingInsideArea() {
+//        timerManager.stopTimer()
+//    }
     
     func setCompletionToLocationManagerDelegate(completion: @escaping (CLLocation) -> ()) {
         let delegate = locationServicesManager.locationManager.delegate
